# docs/openapi-fhir.yaml
openapi: 3.1.0
info:
  title: PCM FHIR API
  version: "2026-01-05"
  description: |
    OpenAPI definition for the PCM FHIR server. This covers the external FHIR REST API
    interactions described in pcm_process_flow.md, including organization onboarding
    updates, service registration, consent creation and retrieval, approvals, and
    endpoint registration. Examples mirror the FHIR examples under input/fsh/_Examples.
    Unless otherwise noted, FHIR REST requests/responses use application/fhir+json.
servers:
  - url: https://pcm.fhir.health.gov.il/r4
    description: PCM FHIR base URL
...
security:
  - mutualTLS: []
  - bearerAuth: []
paths:
...
  /Organization:
    get:
      tags: [Organization]
      summary: Search organizations
      description: |
        Search organizations by supported parameters. Used for discovery and retrieval,
        including searches by active flag or identifier.
      parameters:
        - $ref: '#/components/parameters/_id'
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/active'
        - $ref: '#/components/parameters/type'
        - $ref: '#/components/parameters/partof'
        - $ref: '#/components/parameters/_include'
        - $ref: '#/components/parameters/_include_iterate'
      responses:
        '200':
          description: Bundle of matching organizations
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Bundle'
              examples:
                organizationSearch:
                  $ref: '#/components/examples/OrganizationSearchBundle'
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
  /Organization/{id}:
    get:
      tags: [Organization]
      summary: Read organization
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Organization resource
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Organization'
              examples:
                parentOrganization:
                  $ref: '#/components/examples/OrganizationParentRetrieval'
                serviceProviderOrganization:
                  $ref: '#/components/examples/OrganizationServiceProviderRetrieval'
                dataSourceOrganization:
                  $ref: '#/components/examples/OrganizationDataSourceRetrieval'
                pcmSystemOrganization:
                  $ref: '#/components/examples/OrganizationPcmSystem'
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        '404':
          $ref: '#/components/responses/OperationOutcomeNotFound'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
    put:
      tags: [Organization]
      summary: Update organization
      description: |
        Used by organizations to update allowed fields (contacts, endpoint reference,
        applicable certificate thumbprint (x5t#S256), and redirect URI depending on org type). PCM controls activation.
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Organization'
            examples:
              parentOrganizationUpdate:
                $ref: '#/components/examples/OrganizationParentSubmit'
              serviceProviderUpdate:
                $ref: '#/components/examples/OrganizationServiceProviderSubmit'
              dataSourceUpdate:
                $ref: '#/components/examples/OrganizationDataSourceSubmit'
      responses:
        '200':
          description: Updated organization
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Organization'
              examples:
                parentOrganization:
                  $ref: '#/components/examples/OrganizationParentRetrieval'
                serviceProviderOrganization:
                  $ref: '#/components/examples/OrganizationServiceProviderRetrieval'
                dataSourceOrganization:
                  $ref: '#/components/examples/OrganizationDataSourceRetrieval'
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        '404':
          $ref: '#/components/responses/OperationOutcomeNotFound'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
...
  /Consent:
    get:
      tags: [Consent]
      operationId: searchConsents
      summary: Search consents
      description: |
        Search consents by supported parameters. Use _include to return actor
        organizations, endpoints, and parent orgs in a single bundle.
      parameters:
        - $ref: '#/components/parameters/_id'
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/patient'
        - $ref: '#/components/parameters/pcm_service'
        - $ref: '#/components/parameters/_include'
        - $ref: '#/components/parameters/_include_iterate'
      responses:
        '200':
          description: Bundle of consents
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Bundle'
              examples:
                consentIncludeBundle:
                  $ref: '#/components/examples/ConsentIncludeBundle'
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
    post:
      tags: [Consent]
      operationId: createConsent
      summary: Create consent request
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Consent'
            examples:
              consentRequest:
                $ref: '#/components/examples/ConsentRequest'
      responses:
        '201':
          description: Created consent
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Consent'
              examples:
                consentCreated:
                  $ref: '#/components/examples/ConsentRequestCreated'
          links:
            ReadConsentById:
              operationId: readConsent
              parameters:
                id: '$response.body#/id'
              description: Follow-up read of the created Consent.
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
...
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
      description: mTLS client authentication.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: OAuth2 bearer token (opaque) required for PCM FHIR API access.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
    _id:
      name: _id
      in: query
      required: false
      schema:
        type: string
    identifier:
      name: identifier
      in: query
      required: false
      schema:
        type: string
      description: FHIR identifier search parameter.
...
    active:
      name: active
      in: query
      required: false
      schema:
        type: string
      description: Search by activation status (token; true|false).
...
    status:
      name: status
      in: query
      required: false
      schema:
        type: string
    type:
      name: type
      in: query
      required: false
      schema:
        type: string
      description: Search by type code (Organization.type or HealthcareService.type).
...
    partof:
      name: partof
      in: query
      required: false
      schema:
        type: string
    patient:
      name: patient
      in: query
      required: false
      schema:
        type: string
      description: Search by patient logical reference; supports patient.identifier chain (system|value).
    pcm_service:
      name: pcm-service
      in: query
      required: false
      schema:
        type: string
      description: Search by pcmService extension reference (HealthcareService instance).
    _include:
      name: _include
      in: query
      required: false
      schema:
        type: string
      description: Include related resources (FHIR _include).
    _include_iterate:
      name: _include:iterate
      in: query
      required: false
      schema:
        type: string
      description: Recursive include for related resources.
  responses:
    OperationOutcomeBadRequest:
      description: Bad request (OperationOutcome)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
          examples:
            badRequest:
              $ref: '#/components/examples/OperationOutcomeBadRequest'
    OperationOutcomeUnauthorized:
      description: Unauthorized (OperationOutcome)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
          examples:
            unauthorized:
              $ref: '#/components/examples/OperationOutcomeUnauthorized'
    OperationOutcomeForbidden:
      description: Forbidden (OperationOutcome)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
          examples:
            forbidden:
              $ref: '#/components/examples/OperationOutcomeForbidden'
    OperationOutcomeNotFound:
      description: Not found (OperationOutcome)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
          examples:
            notFound:
              $ref: '#/components/examples/OperationOutcomeNotFound'
    OperationOutcomeError:
      description: Server error (OperationOutcome)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
          examples:
            internalError:
              $ref: '#/components/examples/OperationOutcomeInternalError'
  schemas:
    BaseResource:
      type: object
      required: [resourceType]
      properties:
        resourceType:
          type: string
        id:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
    DomainResource:
      allOf:
        - $ref: '#/components/schemas/BaseResource'
        - type: object
          properties:
            text:
              $ref: '#/components/schemas/Narrative'
            extension:
              type: array
              items:
                $ref: '#/components/schemas/Extension'
...
    Bundle:
      allOf:
        - $ref: '#/components/schemas/BaseResource'
        - type: object
          required: [resourceType, type]
          properties:
            resourceType:
              type: string
              enum: [Bundle]
            type:
              type: string
              enum:
                - document
                - message
                - transaction
                - transaction-response
                - batch
                - batch-response
                - history
                - searchset
                - collection
            total:
              type: integer
            link:
              type: array
              items:
                $ref: '#/components/schemas/BundleLink'
            entry:
              type: array
              items:
                $ref: '#/components/schemas/BundleEntry'
    BundleLink:
      type: object
      properties:
        relation:
          type: string
        url:
          type: string
    BundleEntry:
      type: object
      properties:
        resource:
          $ref: '#/components/schemas/Resource'
        search:
          type: object
          properties:
            mode:
              type: string
    Organization:
      allOf:
        - $ref: '#/components/schemas/DomainResource'
        - type: object
          required: [resourceType]
          properties:
            resourceType:
              type: string
              enum: [Organization]
            active:
              type: boolean
            name:
              type: string
            identifier:
              type: array
              items:
                $ref: '#/components/schemas/Identifier'
            type:
              type: array
              items:
                $ref: '#/components/schemas/CodeableConcept'
            partOf:
              $ref: '#/components/schemas/Reference'
            telecom:
              type: array
              items:
                $ref: '#/components/schemas/ContactPoint'
            address:
              type: array
              items:
                $ref: '#/components/schemas/Address'
            contact:
              type: array
              items:
                $ref: '#/components/schemas/OrganizationContact'
            endpoint:
              type: array
              items:
                $ref: '#/components/schemas/Reference'
...
    Consent:
      allOf:
        - $ref: '#/components/schemas/DomainResource'
        - type: object
          required: [resourceType]
          properties:
            resourceType:
              type: string
              enum: [Consent]
            status:
              type: string
              enum:
                - proposed
                - active
                - rejected
                - inactive
            dateTime:
              type: string
              format: date-time
            identifier:
              type: array
              items:
                $ref: '#/components/schemas/Identifier'
            scope:
              $ref: '#/components/schemas/CodeableConcept'
            category:
              type: array
              items:
                $ref: '#/components/schemas/CodeableConcept'
            patient:
              $ref: '#/components/schemas/Reference'
            provision:
              $ref: '#/components/schemas/ConsentProvision'
...
    Identifier:
      type: object
      properties:
        use:
          type: string
        system:
          type: string
        value:
          type: string
...
    OperationOutcome:
      allOf:
        - $ref: '#/components/schemas/DomainResource'
        - type: object
          required: [resourceType]
          properties:
            resourceType:
              type: string
              enum: [OperationOutcome]
            issue:
              type: array
              items:
                $ref: '#/components/schemas/OperationOutcomeIssue'
  examples:
...
    OrganizationParentSubmit:
      summary: HMO B parent org (submit/update)
      value:
        resourceType: Organization
        active: false
        name: "HMO B (Parent Legal Entity)"
        identifier:
          - use: official
            system: "http://pcm.fhir.health.gov.il/identifier/example-organization-id"
            value: "HMO-B-PARENT-001"
        type:
          - coding:
              - system: "http://fhir.health.gov.il/cs/pcm-org-type"
                code: parent-org
                display: "Parent Organization"
          - coding:
              - system: "http://fhir.health.gov.il/cs/institution-type-moh"
                code: 73
                display: "Health Maintenance Organization"
        telecom:
          - system: phone
            value: "+972-3-555-0200"
        address:
          - text: "10 HMO B Ave, Tel Aviv"
        contact:
          - purpose:
              coding:
                - system: "http://terminology.hl7.org/CodeSystem/contactentity-type"
                  code: ADMIN
            name:
              text: "HMO B - Administration"
            telecom:
              - system: email
                value: "admin@hmo-b.example.org"
          - purpose:
              coding:
                - system: "http://terminology.hl7.org/CodeSystem/contactentity-type"
                  code: PATINF
            name:
              text: "HMO B - Patient Information"
            telecom:
              - system: phone
                value: "+972-3-555-0210"
          - purpose:
              coding:
                - system: "http://fhir.health.gov.il/cs/pcm-contactentity-type"
                  code: SEC
            name:
              text: "HMO B - Security"
            telecom:
              - system: email
                value: "soc@hmo-b.example.org"
    OrganizationParentRetrieval:
      summary: Hospital A parent org (retrieval)
      value:
        resourceType: Organization
        id: org-parent-hospital-a
        meta:
          lastUpdated: "2025-01-01T13:28:17Z"
        active: true
        name: "Hospital A (Parent Legal Entity)"
        identifier:
          - use: official
            system: "http://pcm.fhir.health.gov.il/identifier/example-organization-id"
            value: "HOSP-A-PARENT-001"
        type:
          - coding:
              - system: "http://fhir.health.gov.il/cs/pcm-org-type"
                code: parent-org
                display: "Parent Organization"
          - coding:
              - system: "http://fhir.health.gov.il/cs/institution-type-moh"
                code: 1
                display: "General Hospital"
        telecom:
          - system: phone
            value: "+972-3-555-0100"
        address:
          - text: "1 Hospital A Rd, Jerusalem"
        contact:
          - purpose:
              coding:
                - system: "http://terminology.hl7.org/CodeSystem/contactentity-type"
                  code: ADMIN
            name:
              text: "Hospital A - Administration"
            telecom:
              - system: email
                value: "admin@hospital-a.example.org"
          - purpose:
              coding:
                - system: "http://terminology.hl7.org/CodeSystem/contactentity-type"
                  code: PATINF
            name:
              text: "Hospital A - Patient Information"
            telecom:
              - system: phone
                value: "+972-3-555-0101"
          - purpose:
              coding:
                - system: "http://fhir.health.gov.il/cs/pcm-contactentity-type"
                  code: SEC
            name:
              text: "Hospital A - Security"
            telecom:
              - system: email
                value: "soc@hospital-a.example.org"
    OrganizationServiceProviderSubmit:
      summary: Hospital A service provider (submit)
      value:
        resourceType: Organization
        active: false
        name: "Hospital A - Patient App (Service Provider)"
        identifier:
          - use: official
            system: "http://pcm.fhir.health.gov.il/identifier/example-organization-id"
            value: "HOSP-A-SP-001"
        type:
          - coding:
              - system: "http://fhir.health.gov.il/cs/pcm-org-type"
                code: service-provider
                display: "Service Provider"
          - coding:
              - system: "http://fhir.health.gov.il/cs/institution-type-moh"
                code: 1
                display: "General Hospital"
        partOf:
          reference: "Organization/org-parent-hospital-a"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-oauth2-redirect-uri"
            valueUrl: "https://app.hospital-a.example.org/pcm/oauth2/callback"
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-applicable-certificates"
            extension:
              - url: "thumbprint"
                valueString: "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
    OrganizationServiceProviderRetrieval:
      summary: Hospital A service provider (retrieval)
      value:
        resourceType: Organization
        id: org-sp-hospital-a
        meta:
          lastUpdated: "2025-01-01T13:28:17Z"
        active: true
        name: "Hospital A - Patient App (Service Provider)"
        identifier:
          - use: official
            system: "http://pcm.fhir.health.gov.il/identifier/example-organization-id"
            value: "HOSP-A-SP-001"
        type:
          - coding:
              - system: "http://fhir.health.gov.il/cs/pcm-org-type"
                code: service-provider
                display: "Service Provider"
          - coding:
              - system: "http://fhir.health.gov.il/cs/institution-type-moh"
                code: 1
                display: "General Hospital"
        partOf:
          reference: "Organization/org-parent-hospital-a"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-oauth2-redirect-uri"
            valueUrl: "https://app.hospital-a.example.org/pcm/oauth2/callback"
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-applicable-certificates"
            extension:
              - url: "thumbprint"
                valueString: "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
    OrganizationDataSourceSubmit:
      summary: HMO B data source (submit)
      value:
        resourceType: Organization
        active: false
        name: "HMO B - Clinical Data Source"
        identifier:
          - use: official
            system: "http://pcm.fhir.health.gov.il/identifier/example-organization-id"
            value: "HMO-B-DS-001"
        type:
          - coding:
              - system: "http://fhir.health.gov.il/cs/pcm-org-type"
                code: source
                display: "Data Source"
          - coding:
              - system: "http://fhir.health.gov.il/cs/institution-type-moh"
                code: 73
                display: "Health Maintenance Organization"
        partOf:
          reference: "Organization/org-parent-hmo-b-submit"
        endpoint:
          - reference: "Endpoint/endpoint-hmo-b-submit"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-applicable-certificates"
            extension:
              - url: "thumbprint"
                valueString: "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
    OrganizationDataSourceRetrieval:
      summary: HMO B data source (retrieval)
      value:
        resourceType: Organization
        id: org-ds-hmo-b
        meta:
          lastUpdated: "2025-01-01T13:28:17Z"
        active: true
        name: "HMO B - Clinical Data Source"
        identifier:
          - use: official
            system: "http://pcm.fhir.health.gov.il/identifier/example-organization-id"
            value: "HMO-B-DS-001"
        type:
          - coding:
              - system: "http://fhir.health.gov.il/cs/pcm-org-type"
                code: source
                display: "Data Source"
          - coding:
              - system: "http://fhir.health.gov.il/cs/institution-type-moh"
                code: 73
                display: "Health Maintenance Organization"
        partOf:
          reference: "Organization/org-parent-hmo-b-submit"
        endpoint:
          - reference: "Endpoint/endpoint-hmo-b"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-applicable-certificates"
            extension:
              - url: "thumbprint"
                valueString: "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
    OrganizationPcmSystem:
      summary: PCM system organization
      value:
        resourceType: Organization
        id: org-pcm-system
        meta:
          lastUpdated: "2026-01-05T09:00:00Z"
        active: true
        name: "Patient Consent Manager (PCM)"
        identifier:
          - use: official
            system: "http://institutions.health.gov.il/Institutions"
            value: "PCM"
        type:
          - coding:
              - system: "http://fhir.health.gov.il/cs/pcm-org-type"
                code: pcm
                display: "PCM"
          - coding:
              - system: "http://terminology.hl7.org/CodeSystem/organization-type"
                code: govt
                display: "Government"
        telecom:
          - system: phone
            value: "+972-2-555-0000"
        address:
          - text: "1 Gov Campus Rd, Jerusalem"
        contact:
          - purpose:
              coding:
                - system: "http://terminology.hl7.org/CodeSystem/contactentity-type"
                  code: ADMIN
            name:
              text: "PCM Administration"
            telecom:
              - system: email
                value: "admin@pcm.health.gov.il"
          - purpose:
              coding:
                - system: "http://fhir.health.gov.il/cs/pcm-contactentity-type"
                  code: SEC
            name:
              text: "PCM Security"
            telecom:
              - system: email
                value: "soc@pcm.health.gov.il"
        endpoint:
          - reference: "Endpoint/endpoint-pcm-system"
...
    ConsentRequest:
      summary: Consent request (submit)
      value:
        resourceType: Consent
        status: proposed
        scope:
          coding:
            - system: "http://terminology.hl7.org/CodeSystem/consentscope"
              code: patient-privacy
              display: "Patient Privacy"
        category:
          - coding:
              - system: "http://terminology.hl7.org/CodeSystem/v3-ActCode"
                code: INFA
                display: "information access"
        patient:
          identifier:
            system: "http://fhir.health.gov.il/identifier/il-national-id"
            value: "324573740"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-pcm-service"
            valueReference:
              reference: "HealthcareService/svc-instance-hosp-a-cross-org-summary"
        provision:
          type: permit
          purpose:
            - system: "http://terminology.hl7.org/CodeSystem/v3-ActReason"
              code: TREAT
              display: "treatment"
    ConsentRequestCreated:
      summary: Consent request (created)
      value:
        resourceType: Consent
        id: consent-request-created-hosp-a-cross-org-summary
        meta:
          lastUpdated: "2026-01-05T08:10:00Z"
        identifier:
          - use: official
            system: "http://pcm.fhir.health.gov.il/identifier/pcm-consent-id"
            value: "PCM-CONSENT-0101"
        status: proposed
        dateTime: "2026-01-05T10:10:00+02:00"
        scope:
          coding:
            - system: "http://terminology.hl7.org/CodeSystem/consentscope"
              code: patient-privacy
              display: "Patient Privacy"
        category:
          - coding:
              - system: "http://fhir.health.gov.il/cs/pcm-consent-access-mode"
                code: continuous
                display: "Continuous"
          - coding:
              - system: "http://terminology.hl7.org/CodeSystem/v3-ActCode"
                code: INFA
                display: "information access"
        patient:
          identifier:
            system: "http://fhir.health.gov.il/identifier/il-national-id"
            value: "324573740"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-pcm-service"
            valueReference:
              reference: "HealthcareService/svc-instance-hosp-a-cross-org-summary"
        provision:
          type: permit
          purpose:
            - system: "http://terminology.hl7.org/CodeSystem/v3-ActReason"
              code: TREAT
              display: "treatment"
          actor:
            - role:
                coding:
                  - system: "http://terminology.hl7.org/CodeSystem/v3-ParticipationType"
                    code: IRCP
                    display: "information recipient"
              reference:
                reference: "Organization/org-sp-hospital-a"
...
    OrganizationSearchBundle:
      summary: Organization search bundle
      value:
        resourceType: Bundle
        id: bundle-organization-search
        type: searchset
        total: 2
        entry:
          - search:
              mode: match
            resource:
              resourceType: Organization
              id: org-parent-hospital-a
              name: "Hospital A (Parent Legal Entity)"
          - search:
              mode: match
            resource:
              resourceType: Organization
              id: org-parent-hmo-b-submit
              name: "HMO B (Parent Legal Entity)"
...
    OperationOutcomeBadRequest:
      summary: OperationOutcome - bad request
      value:
        resourceType: OperationOutcome
        id: outcome-bad-request
        issue:
          - severity: error
            code: invalid
            details:
              text: "Invalid request"
            diagnostics: "Missing required field or invalid value."
    OperationOutcomeUnauthorized:
      summary: OperationOutcome - unauthorized
      value:
        resourceType: OperationOutcome
        id: outcome-unauthorized
        issue:
          - severity: error
            code: login
            details:
              text: "Unauthorized"
            diagnostics: "Missing or invalid access token."
    OperationOutcomeForbidden:
      summary: OperationOutcome - forbidden
      value:
        resourceType: OperationOutcome
        id: outcome-forbidden
        issue:
          - severity: error
            code: forbidden
            details:
              text: "Forbidden"
            diagnostics: "Caller not authorized for this resource."
...
    OperationOutcomeInternalError:
      summary: OperationOutcome - internal server error
      value:
        resourceType: OperationOutcome
        id: outcome-internal-error
        issue:
          - severity: error
            code: exception
            details:
              text: "Server error"
            diagnostics: "Unexpected server error."